Added funcitons to execute instructions
Added a few functions tp execute all the instructions in prog1.y86. This
includes creating the registers to store the values and the pointer. The
instructions which are functioning are irmovl, rrmovl, rmmovl, mrmovl, movsbl, cmpl, jmps, writeb, addl, and halt.
Currently the program can execute prog1 and print out the result.